+++ src/main/java/ifpb/factorys/interfaces/IAbstractFactory.java
@@ -1,0 +1,12 @@
+package ifpb.factorys.interfaces;
+
+import ifpb.repositorys.interfaces.IRepositoryAuthors;
+import ifpb.repositorys.interfaces.IRepositoryKeyWords;
+
+public interface IAbstractFactory {
+    IAuthorsFactory createAuthorsFactory();
+    IKeyWorsFactory createKeyWordsFactory();
+    IRepositoryKeyWords createRespositoryKeyWords();
+    IRepositoryAuthors createRepositoryAuthors();
+
+}
