+++ src/main/java/ifpb/factorys/interfaces/IAuthorsFactory.java
@@ -1,13 +1,7 @@
 package ifpb.factorys.interfaces;
 
-import ifpb.entitybasic.interfaces.IAuthor;
+import ifpb.collections.interfaces.IAuthors;
 
 public interface IAuthorsFactory {
-    void add(String matricula,String name);
-    int remove(String matricula);
-    IAuthor get(String matricula);
-    IAuthor[] get();
-
-    int update(String id, String name);
-    void create();
+    IAuthors create();
 }
