+++ src/main/java/ifpb/factorys/interfaces/IKeyWorsFactory.java
@@ -1,15 +1,7 @@
 package ifpb.factorys.interfaces;
 
-import ifpb.entitybasic.exceptions.InvalidNullException;
-import ifpb.entitybasic.interfaces.IKeyWord;
+import ifpb.collections.interfaces.IKeyWords;
 
 public interface IKeyWorsFactory {
-    void add(String keyWord) throws InvalidNullException;
-    void add(String[] keyWordsArray) throws InvalidNullException;
-    int remove(String keyWord) throws InvalidNullException;
-    int update(String oldKeyWord, String newKeyWord) throws InvalidNullException;
-    IKeyWord get(String keyWord) throws InvalidNullException;
-    IKeyWord[] get(String[] keyWords) throws InvalidNullException;
-    IKeyWord[] get();
-    void create()
+    IKeyWords create();
 }
