+++ src/main/java/ifpb/repositorys/interfaces/IRepositoryAuthors.java
@@ -1,0 +1,11 @@
+package ifpb.repositorys.interfaces;
+
+import ifpb.entitybasic.interfaces.IAuthor;
+
+public interface IRepositoryAuthors {
+    void add(String matricula,String name);
+    int remove(String matricula);
+    IAuthor get(String matricula);
+    IAuthor[] get();
+    int update(String id, String name);
+}
